+++ examples/DevEx005_Mixture.jl
@@ -210,7 +210,8 @@ function main(;
         flux = :flux_strided,
         strategy = nothing,
         assembly = :cellwise,
-        npart = 1
+        npart = 1,
+        updatecontrol = true
     )
     h = 1.0 / convert(Float64, n - 1)
     X = collect(0.0:h:1.0)
@@ -262,15 +263,16 @@ function main(;
             strategy.precs.blocksize = nspec
         end
     end
-    control = SolverControl(method_linear = strategy)
+    control = SolverControl(; method_linear = strategy, updatecontrol, damp_initial = 0.5, tol_mono = 1.0e-10)
     control.maxiters = 500
+    @show control.updatecontrol
     u = solve(sys; verbose, control, log = true)
     return norm(u)
 end
 
 using Test
 function runtests()
-
+    verbose = "n"
     if Sys.isapple()
         return nothing
         ## MacOS14 currently crashes here:
@@ -295,11 +297,13 @@ function runtests()
             for flux in [:flux_marray, :flux_strided, :flux_diffcache]
                 for strategy in strategies
                     if dim in strategy.dims
-                        result = main(; dim, assembly, flux, strategy = strategy.method) ≈ dimtestvals[dim]
-                        if !result
-                            @show dim, assembly, flux, strategy
+                        for updatecontrol in [true, false]
+                            result = main(; dim, assembly, flux, strategy = strategy.method, verbose, updatecontrol) ≈ dimtestvals[dim]
+                            if !result
+                                @show dim, assembly, flux, strategy
+                            end
+                            @test result
                         end
-                        @test result
                     end
                 end
             end
@@ -309,8 +313,10 @@ function runtests()
     for dim in [2]
         for assembly in [:edgewise, :cellwise]
             for flux in [:flux_marray, :flux_strided, :flux_diffcache]
-                result = main(; dim, n = 100, assembly, flux, npart = 20)
-                @test  result ≈ 141.54097792523987
+                for updatecontrol in [true, false]
+                    result = main(; dim, n = 100, assembly, flux, npart = 20, verbose, updatecontrol)
+                    @test  result ≈ 141.54097792523987
+                end
             end
         end
     end